# Backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Phone Commerce API
  version: 1.0.0
  description: API documentation for phone e-commerce website
servers:
  - url: https://localhost:3000
    description: Development server
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    get:
      summary: Get all users (admin only)
      tags:
        - Users
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        email:
                          type: string
                        role:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
        '500':
          description: Server error
  /api/users/{id}:
    get:
      summary: Get user by ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      email:
                        type: string
                      role:
                        type: string
                      phone:
                        type: string
                      address:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only access own resource
        '404':
          description: User not found
        '500':
          description: Server error
    put:
      summary: Update user information
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                address:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only update own resource
        '404':
          description: User not found
        '500':
          description: Server error
    delete:
      summary: Delete user
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin or resource owner only
        '404':
          description: User not found
        '500':
          description: Server error

  # Categories endpoints
...
  /api/orders/{id}/checkout:
    post:
      summary: Create a new order (checkout)
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payingMethod
                - shippingMethod
              properties:
                payingMethod:
                  type: string
                  enum: [Cash, 'Credit Card', PayPal]
                  description: Payment method
                shippingMethod:
                  type: string
                  enum: [Standard, Express]
                  description: Shipping method
                addressId:
                  type: integer
                  description: ID of existing address (optional if providing address details)
                fullName:
                  type: string
                  description: Required if not using existing address
                city:
                  type: string
                  description: Required if not using existing address
                district:
                  type: string
                  description: Required if not using existing address
                ward:
                  type: string
                  description: Required if not using existing address
                address:
                  type: string
                  description: Required if not using existing address
                phoneNumber:
                  type: string
                  description: Required if not using existing address
                note:
                  type: string
                state:
                  type: string
                couponId:
                  type: integer
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      orderId:
                        type: integer
        '400':
          description: Invalid input or empty cart
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only create order for own account
        '500':
          description: Server error
  /api/orders/{id}:
    get:
      summary: Get all orders for a user
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: perPage
          schema:
            type: integer
            default: 10
          description: Items per page
      responses:
        '200':
          description: List of user's orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            totalPrice:
                              type: number
                              format: decimal
                            status:
                              type: string
                              enum: [Pending, Delivering, Completed, Cancelled]
                            itemCount:
                              type: integer
                            createdAt:
                              type: string
                              format: date-time
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only access own orders
        '500':
          description: Server error
  /api/orders/{id}/{orderId}:
    get:
      summary: Get details of a specific order
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
          description: Order ID
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      order:
                        type: object
                        properties:
                          id:
                            type: integer
                          userId:
                            type: string
                            format: uuid
                          totalPrice:
                            type: number
                            format: decimal
                          status:
                            type: string
                            enum: [Pending, Delivering, Completed, Cancelled]
                          shippingMethod:
                            type: string
                          payingMethod:
                            type: string
                          address:
                            type: object
                            properties:
                              id:
                                type: integer
                              fullName:
                                type: string
                              city:
                                type: string
                              district:
                                type: string
                              ward:
                                type: string
                              address:
                                type: string
                              phoneNumber:
                                type: string
                              note:
                                type: string
                              state:
                                type: string
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: integer
                                productId:
                                  type: integer
                                productName:
                                  type: string
                                productImage:
                                  type: string
                                quantity:
                                  type: integer
                                price:
                                  type: number
                          createdAt:
                            type: string
                            format: date-time
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only access own orders
        '404':
          description: Order not found
        '500':
          description: Server error
  /api/orders/{id}/{orderId}/status:
    put:
      summary: Update order status
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
          description: Order ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [Pending, Delivering, Completed, Cancelled]
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      orderId:
                        type: integer
                      status:
                        type: string
        '400':
          description: Invalid status
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only update own orders
        '404':
          description: Order not found
        '500':
          description: Server error

  # Images endpoints
...